requirements.txt
Flask==3.1.1
flask_sqlalchemy
flask_migrate
flask_jwt_extended
flask_cors
gunicorn
sqlalchemy
psycopg2
python-dotenv==1.1.1
Werkzeug==3.1.3